Organized code.
