--- apps/pm-user/src/app/pages/index.js
@@ -1,3 +1,0 @@
-import Login from "./login";
-
-export  { Login };
