+++ apps/pm-user/src/assets/index.js
@@ -1,0 +1,1 @@
+export { loginImage } from './loginPage';
