+++ apps/pm-user/src/assets/loginPage/index.js
@@ -1,0 +1,2 @@
+import loginImage from './loginPage.jpg';
+export { loginImage };
