--- apps/pm-user/src/components/formlogin/index.js
@@ -1,3 +1,0 @@
-// import FormLogin from "./FormLogin";
-// export default FormLogin;
-export { FormLogin } from './FormLogin';
