+++ apps/pm-user/src/components/index.js
@@ -1,5 +1,5 @@
-import Button from './button';
-import FormLogin from './formlogin';
-import MyInput from './myinput';
+import Button from './atoms/button';
+import LoginForm from './organisms/loginForm';
+import MyInput from './molecules/myInput';
 
-export { Button, FormLogin, MyInput };
+export { Button, LoginForm, MyInput };
