--- apps/pm-user/src/components/input/Input.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import { FormControl, InputLabel, Input, FormHelperText } from '@mui/material';
-
-const Input = ({ onChange }) => {
-  return (
-    <>
-      {/* <Input id="my-input" aria-describedby="my-helper-text" onChange={} /> */}
-      {/* <Input id="my-input" aria-describedby="my-helper-text" onChange={} /> */}
-    </>
-  );
-};
-
-export default Input;
