--- apps/pm-user/src/components/input/index.js
@@ -1,2 +1,0 @@
-import Input from "./Input";
-export default Input;
