+++ apps/pm-user/src/components/organisms/loginForm/index.js
@@ -1,4 +1,2 @@
 import LoginForm from './LoginForm';
-
-export { LoginForm };
-
+export default LoginForm;
