+++ apps/pm-user/src/pages/home/Home.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import { useOktaAuth } from '@okta/okta-react';
 
 const Home = () => {
