+++ apps/pm-user/src/pages/index.js
@@ -1,0 +1,5 @@
+import Login from './login';
+import Home from './home';
+import Protected from './protected';
+
+export { Login, Home, Protected };
