+++ apps/pm-user/src/pages/protected/index.js
@@ -1,0 +1,2 @@
+import Protected from './Protected';
+export default Protected;
